requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
openai>=1.55.0
pydantic>=2.7.0
pypdf>=5.0.0
python-docx>=1.1.2
Pillow>=10.4.0
pillow-heif>=0.18.0
